frontend_gui/pyproject.toml
[project]
name = "frontend-gui"
version = "0.1.0"
description = "Independent desktop GUI for AiDiy"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mss>=10.1.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.1.0",
    "pyside6>=6.9.0",
    "sounddevice>=0.5.2",
    "websocket-client>=1.9.0",
]

[tool.uv]
package = false
